Time: 47 ms (58.67%), Space: 9.4 MB (72.47%) - LeetHub

+++ 3751-total-waviness-of-numbers-in-range-i/3751-total-waviness-of-numbers-in-range-i.cpp
@@ -5,7 +5,7 @@ public:
         for(int i=num1;i<=num2;i++){
             wave=0;
             string s=to_string(i);
-            for(int j=1;j<s.size()-1;j++){
+            for(size_t j=1;j<s.size()-1;j++){
                 if((s[j]>s[j-1]&&s[j]>s[j+1])||(s[j]<s[j-1]&&s[j]<s[j+1]))wave++;
             }
             sum+=wave;
